Update the application with user roles
