+++ src/popup/components/BrowserPopup.tsx
@@ -19,8 +19,14 @@ import AnnotationSummary from './annotationSummary/AnnotationSummary';
 import { PopupPages } from './BrowserPopupNavigator';
 import '../css/popup.scss';
 import LogoutPanel from './LogoutPanel';
+import { selectUser } from '../../common/store/storage/selectors';
+import { IUserState } from '../../common/store/storage/reducers';
+import { connect } from 'react-redux';
+import { UserRoles } from '../../common/api/user';
 
 export interface IBrowserPopupProps {
+  user: IUserState;
+
   onAnnotationRequestSelect: () => void;
   onPageChange: (PopupPages) => void;
 }
@@ -34,6 +40,15 @@ interface IBrowserPopupState {
   disabledPages: string[];
 }
 
+/*
+ * Currently popup uses data both directly from chrome storage and from redux store
+ * TODO: only use data from redux store
+ */
+@connect(
+  state => ({
+    user: selectUser(state),
+  }),
+)
 export default class BrowserPopup extends React.Component<Partial<IBrowserPopupProps>,
   Partial<IBrowserPopupState>> {
   constructor(props: {}) {
@@ -136,7 +151,7 @@ export default class BrowserPopup extends React.Component<Partial<IBrowserPopupP
       this.setState({ annotationModePages: newAnnotationModePages, requestModePages: newRequestModePages });
       chrome.storage.local.set({
         [chromeKeys.ANNOTATION_MODE_PAGES]: newAnnotationModePages,
-        [chromeKeys.REQUEST_MODE_PAGES]: newRequestModePages
+        [chromeKeys.REQUEST_MODE_PAGES]: newRequestModePages,
       });
       window.close();
       ppGa.annotationAddingModeInited({ location: currentStandardizedTabUrl });
@@ -205,6 +220,55 @@ export default class BrowserPopup extends React.Component<Partial<IBrowserPopupP
     ppGa.extensionReportButtonClicked({ location: this.state.currentStandardizedTabUrl });
   }
 
+  renderFeatureButtons() {
+    const {
+      isExtensionDisabled,
+    } = this.state;
+
+    const isCurrentPageDisabled = this.isExtensionDisabledForCurrentTab();
+    const isAnnotationMode = this.isAnnotationModeForCurrentTab();
+    const isRequestMode = this.isRequestModeForCurrentTab();
+
+    return (
+      <>
+        {this.props.user.userRole === UserRoles.editor &&
+        <li
+          className={classNames('menu-item', 'clickable', 'primary',
+            { disabled: isExtensionDisabled || isCurrentPageDisabled },
+            { active: isAnnotationMode })}
+          onClick={this.handleAnnotationModeClick}
+        >
+          <Icon className="icon" icon={ic_add_circle} size={25}/>
+          {isAnnotationMode ?
+            <span>Dodajesz przypis </span>
+            : <span>Dodaj przypis</span>
+          }
+          <p className="caption">
+            Dodaj przypis, aby podzielić się wartościową informacją ze społecznością *PP
+          </p>
+        </li>
+        }
+
+        <li
+          className={classNames('menu-item', 'clickable',
+            { disabled: isExtensionDisabled || isCurrentPageDisabled },
+            { active: isRequestMode })}
+          onClick={this.handleAnnotationRequestClick}
+        >
+          <Icon className="icon" icon={ic_live_help} size={25}/>
+          {isRequestMode ?
+            <span>Zgłaszasz prośbę o przypis </span>
+            : <span>Poproś o przypis</span>
+          }
+          <p className="caption">Możesz poprosić o sprawdzenie wybranego fragmentu artykułu.
+            Twoje zgłoszenie zostanie przekazane do redaktorów Demagoga.
+          </p>
+        </li>
+        <hr className="menu-separator"/>
+      </>
+    );
+  }
+
   render() {
     const {
       isLoading,
@@ -212,8 +276,6 @@ export default class BrowserPopup extends React.Component<Partial<IBrowserPopupP
     } = this.state;
 
     const isCurrentPageDisabled = this.isExtensionDisabledForCurrentTab();
-    const isAnnotationMode = this.isAnnotationModeForCurrentTab();
-    const isRequestMode = this.isRequestModeForCurrentTab();
 
     if (isLoading) {
       // Do not display the page already, when loading; otherwise we'll always see the toggle transition...
@@ -232,37 +294,9 @@ export default class BrowserPopup extends React.Component<Partial<IBrowserPopupP
               </a>
             </div>
             <AnnotationSummary onFullViewClick={this.handleFullAnnotationViewClick}/>
-            <li
-              className={classNames('menu-item', 'clickable', 'primary',
-                { disabled: isExtensionDisabled || isCurrentPageDisabled },
-                { active: isAnnotationMode })}
-              onClick={this.handleAnnotationModeClick}
-            >
-              <Icon className="icon" icon={ic_add_circle} size={25}/>
-              {isAnnotationMode ?
-                <span>Dodajesz przypis </span>
-                : <span>Dodaj przypis</span>
-              }
-              <p className="caption">
-                Dodaj przypis, aby podzielić się wartościową informacją ze społecznością *PP
-              </p>
-            </li>
-            <li
-              className={classNames('menu-item', 'clickable',
-                { disabled: isExtensionDisabled || isCurrentPageDisabled },
-                { active: isRequestMode })}
-              onClick={this.handleAnnotationRequestClick}
-            >
-              <Icon className="icon" icon={ic_live_help} size={25}/>
-              {isRequestMode ?
-                <span>Zgłaszasz prośbę o przypis </span>
-                : <span>Poproś o przypis</span>
-              }
-              <p className="caption">Możesz poprosić o sprawdzenie wybranego fragmentu artykułu.
-                Twoje zgłoszenie zostanie przekazane do redaktorów Demagoga.
-              </p>
-            </li>
-            <hr className="menu-separator"/>
+
+            {this.renderFeatureButtons()}
+
             <li className="menu-item">
               <Icon className="icon" icon={ic_block} size={25}/>
               <span>Wyłącz przypisy</span>
